requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
jinja2==3.1.4
sqlalchemy==2.0.32
pydantic==2.8.2
python-multipart==0.0.9
qrcode==7.4.2
pillow==10.4.0
openpyxl==3.1.5
reportlab==4.2.2
python-dotenv==1.0.1
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9

psycopg2-binary==2.9.9

pyotp
python-json-logger
qrcode[pil]
